Move updateItem logic to firebase API and handle cases when required dates are null
